==> Data processing/05 - hydrolakes.Rmd <==
---
title: "Hydrolakes"
author: "Abby Lewis"
date: "2023-03-23"
output: html_document
---

This file compiles hydrolakes metadata using unique hydrolakes IDs. The raw hydrolakes shapefile is too large to store on GitHub, so will need to be re-downloaded. If you would rather skip this step, the output of this code file (hydrolakes metadata for each lake) has been created and stored on github in "./External data/hydrolakes_full.csv"

To download raw hydrolakes data: 
https://www.hydrosheds.org/products/hydrolakes
  Lake pour points (including all attributes) in shapefileformat (79MB).

Table of contents:
Step 1: Load packages
Step 2: Add hydrolakes data


Step 1: Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
#install.packages("foreign")
library(foreign)
```


Step 2: Add hydrolakes data
```{r}
dbf <- read.dbf("../External data/HydroLAKES_points_v10_shp/HydroLAKES_points_v10.dbf")

ids <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")%>%
  select(LakeID,hylak_id)%>%
  rename(ID = LakeID)

comb <- ids%>%
  left_join(dbf, by = c("hylak_id"="Hylak_id"))

write.csv(comb,"../Compiled data/hydrolakes_full.csv")
```
