==> Data processing/08 - Anoxic factor.Rmd <==
---
title: "ASL AF calcs"
author: "Abby Lewis"
date: "2022-08-29"
output: html_document
---

This file calculates anoxic factor for each lake-year, treating the entire hypolimnion as one layer. 

Step 1: Load data and packages
Step 2: Calculate VW DO
Step 3: Model end of stratification date
Step 4: Model start of anoxia date using oxygen demand
Step 5: Empirically calculate start of anoxia date if it can be determined within a certain range (PROXIMITY)
Step 6: Calculate AF, combining all info
Step 7: If VW DO is high all summer, AF = 0 (filling in lakes where we couldn't use oxygen demand)
Step 8: Export and plot


Step 1: Load data and packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
library(openair)
library(ggpubr)

#Load data and stratification boundaries. This file is created by "03 - Stratified avgs.Rmd"
full_with_thermo <- read.csv("../Compiled data/Stratified_period_data_with_thermo.csv")%>%
  mutate(Date=as.Date(Date))
#Load lake metadata
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
#Load compiled data
annual_data <- read.csv("../Compiled data/All_data_annual.csv")
```


Step 2: Calculate VW DO
```{r}
#prepare bathymetry
lake_bats <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/ba733454c29a5f026efed05a6e6ef75b")%>%
  filter(Area_m2>0)%>%
  group_by(LakeID)%>%
  mutate(methods_n = length(unique(Method)))%>%
  filter(methods_n==1|Method=="data")%>%
  ungroup()

#Calculate volume-weighted oxygen concentrations
vw_do <- full_with_thermo%>%
  full_join(lake_bats, by = c("LakeID","Depth_m"))%>%
  filter(!is.na(IntervalVolume_m3))%>%
  filter(Depth_m>hypo_depth)%>%
  mutate(DO_mass = DO_mgL*IntervalVolume_m3,
         Temp_total = Temp_C*IntervalVolume_m3,
         Year = year(Date))%>%
  group_by(Date, LakeID)%>%
  dplyr::summarize(DO_tot = sum(DO_mass),#sum across all hypolimnetic depths
                   vol_tot = sum(IntervalVolume_m3),
                   DO_mgL_vol = DO_tot/vol_tot,#Divide by hypolimnetic volume
                   DO_mgL_area = DO_tot/max(Area_m2),
                   Temp_C = sum(Temp_total)/vol_tot)%>%
  mutate(Year= year(Date))%>%
  group_by(LakeID, Year)%>%
  filter(!is.na(DO_mgL_vol))%>%
  arrange(LakeID, Date)%>%
  mutate(low_point = ifelse(sum(DO_mgL_vol<1)==0,#Identify the lowest DO value for a given year
                            Date[which.min(DO_mgL_vol)],
                            first(Date[DO_mgL_vol<1])))%>%
  filter(is.na(low_point)|Date<=low_point)%>% #Remove days after the lowest DO value
  ungroup()%>%
  mutate(Year = year(Date))
```


Step 3: Model end of stratification date
```{r}
# Calculate strat end if it can be determined within a certain range (PROXIMITY)
PROXIMITY = 7

empirical_strat_end <- full_with_thermo%>%
  full_join(lake_bats, by = c("LakeID","Depth_m"))%>%
  group_by(Date, LakeID)%>%
  dplyr::summarize(mixing_date = unique(Date_end)) %>%
  mutate(Year= year(Date))%>%
  group_by(LakeID, Year)%>%
  filter(Date == last(Date[Date <= as.Date(mixing_date)])) %>%
  mutate(strat_interval = as.numeric(as.Date(mixing_date) - Date),
         mixing_date = yday(mixing_date) - (strat_interval/2)) %>%
  filter(strat_interval <= PROXIMITY) %>%
  select(LakeID, Year, mixing_date)

#Nürnberg model
strat_end <- annual_data %>%
  mutate(strat_end_model = 10^(2.62 + 
                           (-0.116 * log10(Temp_C_HYPO)) + 
                           (0.042 * log10(MeanDepth_m)) + 
                           (-0.002 * Latitude_DD)),
         strat_end_model_corrected = strat_end_model + 2.5/10 * (Year - 1988)) #correction from Jane et al. 2023

#New model incorporating year
for_new_model <- annual_data %>%
  full_join(empirical_strat_end) %>%
  dplyr::select(mixing_date, Temp_C_HYPO, MeanDepth_m, Latitude_DD, Year, LakeID) %>%
  drop_na()

length(unique(for_new_model$LakeID))
nrow(for_new_model)

new_model <- lm(log10(for_new_model$mixing_date) ~ log10(for_new_model$Temp_C_HYPO) +
     log10(for_new_model$MeanDepth_m) +
     for_new_model$Latitude_DD +
     for_new_model$Year)

#Generate predictions
new_strat_end_model <- annual_data %>%
  mutate(strat_end_new_model = 10^(new_model$coefficients[1] + 
                                 (new_model$coefficients[2] * log10(Temp_C_HYPO)) + 
                                 (new_model$coefficients[3] * log10(MeanDepth_m)) + 
                                 (new_model$coefficients[4] * Latitude_DD) +
                                 (new_model$coefficients[5] * Year)
                                 )) %>%
  dplyr::select(LakeID, Year, strat_end_new_model) %>%
  unique()

summary(new_model)
```


Step 4: Model start of anoxia date using oxygen demand
```{r}
#Calculate the rate of change in volume-weighted concentrations using lm
DO_THRESH = 1.8

vw_do_demand <- vw_do%>%
  mutate(Year = year(Date),
         DOY = yday(Date))%>%
  group_by(Year, LakeID)%>%
  dplyr::summarize(n = n(),
            DO_demand_mgLd = -lm(DO_mgL_vol~DOY)$coefficients[2],
            r2_vol = summary(lm(DO_mgL_vol~DOY))$r.squared,
            anoxic_onset = (DO_THRESH - lm(DO_mgL_vol~DOY)$coefficients[1]) / 
              lm(DO_mgL_vol~DOY)$coefficients[2])%>%
  filter(n>=3)#Need at least 3 points

#How many rows get removed by filtering so R2>0.5? 246
removed_by_r2 <- vw_do_demand%>%
  filter(r2_vol>.5)

#How many rows get removed by filtering so oxygen demand is positive? 0
removed_by_pos <- removed_by_r2%>%
  filter(DO_demand_mgLd>0)

#Finish QAQC
good_hod <- vw_do_demand%>%
  group_by(LakeID)%>%
  filter(!is.na(DO_demand_mgLd),
         r2_vol>0.5
         )%>%
  dplyr::summarize(n_neg = sum(DO_demand_mgLd<0),
                   n = n(),
                   pct = n_neg/n*100)%>%
  filter(pct<=10)

vw_do_demand_qaqc <- removed_by_pos%>%
  filter(LakeID %in% good_hod$LakeID)
```


Step 5: Empirically calculate start of anoxia date if it can be determined within a certain range (PROXIMITY)
```{r}
PROXIMITY = 7

empirical_anoxic_onset <- vw_do %>%
  group_by(LakeID, Year)%>%
  filter(sum(DO_mgL_vol < DO_THRESH, na.rm = T) >= 1,
         sum(DO_mgL_vol >= DO_THRESH & 
               Date < first(Date[DO_mgL_vol < DO_THRESH]), 
             na.rm = T) >= 1) %>%
  mutate(onset = first(Date[DO_mgL_vol < DO_THRESH]),
         before = last(Date[Date<onset]),
         interval = as.numeric(onset-before),
         range = DO_mgL_vol[Date==before] - DO_mgL_vol[Date==onset],
         fraction = (DO_mgL_vol[Date==before] - DO_THRESH) / range)%>%
  filter(interval <= PROXIMITY) %>%
  summarize(anoxic_onset_data = unique(yday(before) + fraction * interval))
```


Step 6: Calculate AF, combining all info
```{r}
anoxic_on_off = vw_do_demand_qaqc %>%
  dplyr::select(Year, LakeID, anoxic_onset) %>%
  rename(anoxic_onset_vhod = anoxic_onset) %>%
  full_join(empirical_anoxic_onset) %>%
  full_join(empirical_strat_end) %>%
  full_join(new_strat_end_model) %>%
  full_join(strat_end %>% dplyr::select(LakeID, Year, strat_end_model, strat_end_model_corrected))

lake_bats <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/ba733454c29a5f026efed05a6e6ef75b")%>%
  filter(Area_m2>0)%>%
  group_by(LakeID)%>%
  mutate(methods_n = length(unique(Method)))%>%
  filter(methods_n==1|Method=="data")%>%
  ungroup()

areas = full_with_thermo %>%
  group_by(LakeID, Year) %>%
  summarize(hypo_depth = round(mean(hypo_depth)),
            SurfaceArea_ha = unique(SurfaceArea_ha)) %>%
  left_join(lake_bats%>% select(LakeID, Depth_m, Area_m2), by = c("LakeID",hypo_depth = "Depth_m"))

AF = anoxic_on_off %>%
  mutate(anoxic_onset = ifelse(!is.na(anoxic_onset_data),
                               anoxic_onset_data,
                               anoxic_onset_vhod),
         anoxic_end = strat_end_new_model,
         duration = anoxic_end - anoxic_onset) %>%
  left_join(areas) %>%
  mutate(AF = duration*Area_m2/(SurfaceArea_ha*10000),
         AF = ifelse(AF < 0, 0, AF))
```


Step 7: If VW DO is high all summer, AF = 0 (filling in lakes where we couldn't use oxygen demand)
```{r}
summer <- read_csv("../Compiled data/summer_averages_wi.csv")
summer_af <- summer %>%
  group_by(LakeID, Year) %>%
  filter(Layer == "HYPO",
         DO_mgL > 5) %>%
  mutate(AF_summer = 0) %>%
  select(LakeID, Year, AF_summer)
  
AF_output <- AF %>%
  dplyr::select(LakeID, Year, AF) %>%
  full_join(summer_af) %>%
  mutate(AF = ifelse(is.na(AF), AF_summer, AF)) %>%
  dplyr::select(LakeID, Year, AF) %>%
  filter(!is.na(AF))

length(unique(AF_output$LakeID))
nrow(AF_output)
```


Step 8: Export and plot
```{r}
#Save results
write.csv(AF_output, "../Compiled data/AF.csv", row.names = F)

#Plot anoxic onset (data vs. VHOD)
jpeg("../Figures/Anoxic onset comparison.jpg", width = 3.75, height = 3.5, res = 300, units = "in")
anoxic_on_off %>%
  filter(!is.na(anoxic_onset_data),
         !is.na(anoxic_onset_vhod)) %>%
  ggplot(aes(x = anoxic_onset_data, y = anoxic_onset_vhod))+
  #geom_abline(slope = 1)+
  geom_smooth(method = "lm", formula = y~x)+
  geom_point()+
  theme_bw()+
  xlab("Anoxic onset (between profiles)") + 
  ylab("Anoxic onset (VHOD)")+
  stat_regline_equation(aes(label =  paste(after_stat(eq.label), after_stat(adj.rr.label), sep = "~~~~~~~")))
dev.off()

#Plot mixing dates with multiple methods
p1 <- anoxic_on_off %>%
  filter(!is.na(mixing_date), !is.na(strat_end_model)) %>%
  ggplot(aes(x = mixing_date, y = strat_end_model))+
  #geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()+
  xlab("Stratification end (between profiles)") + 
  ylab("Stratification end (Nürnberg 1988)")+
  theme_bw()+
  stat_regline_equation(aes(label =  paste(after_stat(eq.label), after_stat(adj.rr.label), sep = "~~~~~~~")))

p2 <- anoxic_on_off %>%
  filter(!is.na(mixing_date), !is.na(strat_end_new_model)) %>%
  ggplot(aes(x = mixing_date, y = strat_end_new_model))+
  #geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()+
  xlab("Stratification end (between profiles)") + 
  ylab("Stratification end (modified model)")+
  theme_bw()+
  stat_regline_equation(aes(label =  paste(after_stat(eq.label), after_stat(adj.rr.label), sep = "~~~~~~~")))

jpeg("../Figures/Strat end comparison.jpg", width = 6.5, height = 3.5, res = 300, units = "in")
ggarrange(p1,p2)
dev.off()

#Summary statistics
summary(lm(anoxic_on_off$mixing_date ~ anoxic_on_off$strat_end_model))
summary(lm(anoxic_on_off$mixing_date ~ anoxic_on_off$strat_end_new_model))

#Add to compiled data
with_temp_af <- annual_data %>%
  left_join(AF_output)

write.csv(with_temp_af,"../Compiled data/All_data_annual AF.csv",row.names =F)
```
